Set up machanism for choosing data store
